requirements.txt
Flask==3.0.3
flask-cors==4.0.0
openai==1.14.3
python-dotenv==1.0.1
